mw-puppet-client: inherit puppet::client::cron
We need to use either cron or daemon for puppet to run automatically. Given a
a number of discussions I've found, cron still seems to be the best solution.
People are still having trouble with the daemon and the running interval is
still hardcoded to 30 minutes. That's why I propose to carry on with the cron
job for the time being.