generic-tmpl: renamed os-debian-squeeze-locales to os::locales
This means this class will have to work on all distribs and versions. This has
only been tested on squeeze so far, but it doesn't seem to have any specific
recipe, so it might already be generic as is.

Not sure if we should have ifs with 'fail' for untested distribs?